Support plot filter change for testnet10
