+++ packages/api/src/constants/PlotFilter.ts
@@ -1,0 +1,16 @@
+const PLOT_FILTER_CONSTANTS = {
+  mainnet: {
+    HARD_FORK_HEIGHT: 5_496_000,
+    PLOT_FILTER_128_HEIGHT: 10_542_000,
+    PLOT_FILTER_64_HEIGHT: 15_592_000,
+    PLOT_FILTER_32_HEIGHT: 20_643_000,
+  },
+  testnet10: {
+    HARD_FORK_HEIGHT: 2_997_292,
+    PLOT_FILTER_128_HEIGHT: 3_061_804,
+    PLOT_FILTER_64_HEIGHT: 8_010_796,
+    PLOT_FILTER_32_HEIGHT: 13_056_556,
+  },
+};
+
+export default PLOT_FILTER_CONSTANTS;
